+++ hpc-engine/cpca.c
@@ -799,9 +799,13 @@ int main( int argc , char ** argv )
     // densities - local reduction
     gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
     // densities - global reduction
+#ifdef USE_GPU_AWARE_MPI
+    MPI_Allreduce( MPI_IN_PLACE, d_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
     gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)12*NXY*sizeof(double)) ); 
     MPI_Allreduce( MPI_IN_PLACE, h_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
     gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)12*NXY*sizeof(double)) ); 
+#endif
     if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
     if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);
@@ -962,9 +966,13 @@ int main( int argc , char ** argv )
             // densities - local reduction
             gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
+#ifdef USE_GPU_AWARE_MPI
+            MPI_Allreduce( MPI_IN_PLACE, d_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
             gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)12*NXY*sizeof(double)) ); 
             MPI_Allreduce( MPI_IN_PLACE, h_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
             gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)12*NXY*sizeof(double)) );
+#endif
             if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
             if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);
@@ -1044,9 +1052,13 @@ int main( int argc , char ** argv )
             // densities - local reduction
             gpu_exec( calculate_densities(nwfip, d_fkm3, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
+#ifdef USE_GPU_AWARE_MPI
+            MPI_Allreduce( MPI_IN_PLACE, d_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
             gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)12*NXY*sizeof(double)) ); 
             MPI_Allreduce( MPI_IN_PLACE, h_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
             gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)12*NXY*sizeof(double)) );
+#endif
             if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
             if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);
@@ -1144,9 +1156,13 @@ int main( int argc , char ** argv )
         // densities - local reduction
         gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
         // densities - global reduction
+#ifdef USE_GPU_AWARE_MPI
+        MPI_Allreduce( MPI_IN_PLACE, d_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
         gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)12*NXY*sizeof(double)) ); 
         MPI_Allreduce( MPI_IN_PLACE, h_densities, 12*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
         gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)12*NXY*sizeof(double)) ); 
+#endif
         if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
         if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);
@@ -1227,9 +1243,13 @@ int main( int argc , char ** argv )
             gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
             mpipackagesize = EXCHANGE_SIZE;
+#ifdef USE_GPU_AWARE_MPI
+            MPI_Allreduce( MPI_IN_PLACE, d_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
             gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)mpipackagesize*NXY*sizeof(double)) ); 
             MPI_Allreduce( MPI_IN_PLACE, h_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
             gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)mpipackagesize*NXY*sizeof(double)) );
+#endif
             if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
             if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);
@@ -1278,9 +1298,13 @@ int main( int argc , char ** argv )
             gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
             if(i_step==md.timesteps-1) mpipackagesize = 12; else mpipackagesize = EXCHANGE_SIZE;
+#ifdef USE_GPU_AWARE_MPI
+            MPI_Allreduce( MPI_IN_PLACE, d_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+#else
             gpu_exec( memcopy_gpu2host(d_densities, h_densities,  (size_t)mpipackagesize*NXY*sizeof(double)) ); 
             MPI_Allreduce( MPI_IN_PLACE, h_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
             gpu_exec( memcopy_host2gpu(h_densities, d_densities,  (size_t)mpipackagesize*NXY*sizeof(double)) ); 
+#endif
             if(md.spinsymmetry>0) symmetrize_densities_device(d_densities); // special calse: spin-symmetric system
 #ifndef TAU_COMPUTATION_VIA_GRADIENTS
             if(gradients_computed) density_caculate_tau(d_densities, md.nthreads);