+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2022.xx.xx
+* The td codes can boost the performance by exploiting GPU-aware MPI.
+
 # 2022.08.27
 * Porting code to [HIP](https://gitlab.fizyka.pw.edu.pl/wtools/wslda/-/wikis/Converting%20the%20toolkit%20to%20HIP).
 * New library [W-Box](https://gitlab.fizyka.pw.edu.pl/wtools/wbox) has been incorporated into the W-SLDA.